# gen/http/openapi.json
{"swagger":"2.0","info":{"title":"Stress test service for kagome","description":"Service for kagome stress tests","version":""},"host":"localhost:8000","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/openapi.json":{"get":{"summary":"Download ./gen/http/openapi.json","operationId":"stress#/openapi.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"]}},"/start":{"post":{"tags":["stress"],"summary":"start stress","operationId":"stress#start","responses":{"204":{"description":"No Content response."}},"schemes":["http"]}},"/stop":{"post":{"tags":["stress"],"summary":"stop stress","operationId":"stress#stop","responses":{"204":{"description":"No Content response."}},"schemes":["http"]}},"/tokenize":{"post":{"tags":["stress"],"summary":"tokenize stress","operationId":"stress#tokenize","parameters":[{"name":"TokenizeRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/StressTokenizeRequestBody","required":["sentence"]}}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/StressTokenResponseCollection"}}},"schemes":["http"]}}},"definitions":{"StressTokenResponseCollection":{"title":"Mediatype identifier: application/vnd.token+json; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/TokenResponse"},"description":"TokenizeResponseBody is the result type for an array of TokenResponse (default view)","example":[{"end":6937734305957636642,"pos":"Tempore explicabo.","start":5608356791695972864,"surface":"Adipisci qui laudantium pariatur aspernatur.","type":"Dolorum ipsa nisi dolores doloremque maiores."},{"end":6937734305957636642,"pos":"Tempore explicabo.","start":5608356791695972864,"surface":"Adipisci qui laudantium pariatur aspernatur.","type":"Dolorum ipsa nisi dolores doloremque maiores."},{"end":6937734305957636642,"pos":"Tempore explicabo.","start":5608356791695972864,"surface":"Adipisci qui laudantium pariatur aspernatur.","type":"Dolorum ipsa nisi dolores doloremque maiores."},{"end":6937734305957636642,"pos":"Tempore explicabo.","start":5608356791695972864,"surface":"Adipisci qui laudantium pariatur aspernatur.","type":"Dolorum ipsa nisi dolores doloremque maiores."}]},"StressTokenizeRequestBody":{"title":"StressTokenizeRequestBody","type":"object","properties":{"sentence":{"type":"string","example":"Dolore culpa a."}},"example":{"sentence":"Eligendi modi commodi."},"required":["sentence"]},"TokenResponse":{"title":"Mediatype identifier: application/vnd.token+json; view=default","type":"object","properties":{"end":{"type":"integer","example":9138893857102516976,"format":"int64"},"pos":{"type":"string","example":"Sint distinctio suscipit sapiente."},"start":{"type":"integer","example":2599820768220399435,"format":"int64"},"surface":{"type":"string","example":"Nisi error id rerum."},"type":{"type":"string","example":"Eum excepturi sed sint ducimus soluta velit."}},"description":"TokenResponse result type (default view)","example":{"end":8111107198680295956,"pos":"Iste ut aliquid.","start":2832631226233261616,"surface":"Sunt corporis.","type":"Aliquam velit."},"required":["surface","pos","start","end","type"]}}}