# gen/http/stress/server/encode_decode.go
// Code generated by goa v3.0.6, DO NOT EDIT.
//
// stress HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/ikawaha/kagome-stress/design

package server

import (
	"context"
	"io"
	"net/http"

	stressviews "github.com/ikawaha/kagome-stress/gen/stress/views"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeStartResponse returns an encoder for responses returned by the stress
// start endpoint.
func EncodeStartResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		w.WriteHeader(http.StatusNoContent)
		return nil
	}
}

// EncodeStopResponse returns an encoder for responses returned by the stress
// stop endpoint.
func EncodeStopResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		w.WriteHeader(http.StatusNoContent)
		return nil
	}
}

// EncodeTokenizeResponse returns an encoder for responses returned by the
// stress tokenize endpoint.
func EncodeTokenizeResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res := v.(stressviews.TokenCollection)
		enc := encoder(ctx, w)
		body := NewTokenResponseCollection(res.Projected)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeTokenizeRequest returns a decoder for requests sent to the stress
// tokenize endpoint.
func DecodeTokenizeRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (interface{}, error) {
	return func(r *http.Request) (interface{}, error) {
		var (
			body TokenizeRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateTokenizeRequestBody(&body)
		if err != nil {
			return nil, err
		}
		payload := NewTokenizePayload(&body)

		return payload, nil
	}
}